remove "monaco" from name of monaco-field components
These are no longer backed by Monaco; they use CodeMirror.

Also refactor to not use the needless forwardRef.
