+++ client/web/BUILD.bazel
@@ -775,11 +775,11 @@ ts_project(
         "src/enterprise/insights/components/creation-ui/sanitizers/repositories.ts",
         "src/enterprise/insights/components/creation-ui/sanitizers/series.ts",
         "src/enterprise/insights/components/creation-ui/validators/validators.ts",
+        "src/enterprise/insights/components/form/field/Field.tsx",
+        "src/enterprise/insights/components/form/field/PreviewLink.tsx",
+        "src/enterprise/insights/components/form/field/index.ts",
         "src/enterprise/insights/components/form/form-radio-input/FormRadioInput.tsx",
         "src/enterprise/insights/components/form/index.ts",
-        "src/enterprise/insights/components/form/monaco-field/MonacoField.tsx",
-        "src/enterprise/insights/components/form/monaco-field/MonacoPreviewLink.tsx",
-        "src/enterprise/insights/components/form/monaco-field/index.ts",
         "src/enterprise/insights/components/form/query-input/InsightQueryInput.tsx",
         "src/enterprise/insights/components/form/query-input/utils/generate-repo-filters-query.ts",
         "src/enterprise/insights/components/form/repositories-field/RepositoriesField.tsx",
