+++ client/web/src/enterprise/insights/components/creation-ui/insight-repo-section/InsightRepoSection.tsx
@@ -38,8 +38,8 @@ import type {
     InsightRepositoriesCountVariables,
 } from '../../../../../graphql-operations'
 import type { CreateInsightFormFields } from '../../../pages/insights/creation/search-insight'
-import { getRepoQueryPreview, RepositoriesField, MonacoField } from '../../form'
-import { MonacoPreviewLink } from '../../form/monaco-field'
+import { getRepoQueryPreview, RepositoriesField, Field } from '../../form'
+import { PreviewLink } from '../../form/field'
 
 import styles from './InsightRepoSection.module.scss'
 
@@ -251,7 +251,7 @@ function SmartSearchQueryRepoField(props: SmartSearchQueryRepoFieldProps): React
                 )}
 
                 <InputElement
-                    as={MonacoField}
+                    as={Field}
                     queryState={queryState}
                     status={fieldStatus}
                     placeholder="Example: repo:sourcegraph/*"
@@ -264,14 +264,14 @@ function SmartSearchQueryRepoField(props: SmartSearchQueryRepoFieldProps): React
                     {...attributes}
                 />
 
-                <MonacoPreviewLink
+                <PreviewLink
                     query={previewQuery}
                     patternType={SearchPatternType.standard}
                     className={styles.repoLabelPreviewLink}
                     tabIndex={disabled ? -1 : 0}
                 >
                     <LinkExternalIcon size={18} />
-                </MonacoPreviewLink>
+                </PreviewLink>
             </LabelComponent>
 
             <SmartRepoQueryChips disabled={disabled} onChipClick={handleChipSuggestions} />
