+++ client/web/src/enterprise/insights/components/form/index.ts
@@ -6,4 +6,4 @@ export { RepositoryField } from './repositories-field/RepositoryField'
 export { RepositoriesField } from './repositories-field/RepositoriesField'
 export { InsightQueryInput } from './query-input/InsightQueryInput'
 export { FormRadioInput } from './form-radio-input/FormRadioInput'
-export { MonacoField } from './monaco-field'
+export { Field } from './field'
