--- client/web/src/enterprise/insights/components/form/monaco-field/index.ts
@@ -1,18 +1,0 @@
-import { MonacoField, MonacoFocusContainer, type MonacoFieldProps } from './MonacoField'
-import { MonacoPreviewLink } from './MonacoPreviewLink'
-
-const Root = MonacoFocusContainer
-const Field = MonacoField
-const PreviewLink = MonacoPreviewLink
-
-export {
-    MonacoFocusContainer,
-    MonacoField,
-    MonacoPreviewLink,
-    //
-    Root,
-    Field,
-    PreviewLink,
-}
-
-export type { MonacoFieldProps }
