+++ client/web/src/enterprise/insights/components/form/query-input/InsightQueryInput.tsx
@@ -1,4 +1,4 @@
-import { forwardRef, type InputHTMLAttributes, type PropsWithChildren } from 'react'
+import type { InputHTMLAttributes } from 'react'
 
 import classNames from 'classnames'
 import LinkExternalIcon from 'mdi-react/OpenInNewIcon'
@@ -6,17 +6,17 @@ import LinkExternalIcon from 'mdi-react/OpenInNewIcon'
 import type { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 import { QueryChangeSource, type QueryState } from '@sourcegraph/shared/src/search'
 
-import type { MonacoFieldProps } from '../monaco-field'
-import * as Monaco from '../monaco-field'
+import type { FieldProps } from '../field'
+import { Field, FocusContainer, PreviewLink } from '../field'
 
 import { generateRepoFiltersQuery, getRepoQueryPreview } from './utils/generate-repo-filters-query'
 
 import styles from './InsightQueryInput.module.scss'
 
 type NativeInputProps = Omit<InputHTMLAttributes<HTMLInputElement>, 'onChange' | 'onBlur'>
-type MonacoPublicProps = Omit<MonacoFieldProps, 'queryState' | 'onChange' | 'aria-invalid'>
+type FieldPublicProps = Omit<FieldProps, 'queryState' | 'onChange' | 'aria-invalid'>
 
-export interface InsightQueryInputProps extends MonacoPublicProps, NativeInputProps {
+export interface InsightQueryInputProps extends FieldPublicProps, NativeInputProps {
     value: string
     repoQuery: string | null
     repositories: string[]
@@ -24,59 +24,54 @@ export interface InsightQueryInputProps extends MonacoPublicProps, NativeInputPr
     onChange: (value: string) => void
 }
 
-export const InsightQueryInput = forwardRef<HTMLInputElement, PropsWithChildren<InsightQueryInputProps>>(
-    (props, reference) => {
-        const {
-            value,
-            patternType,
-            repoQuery,
-            repositories = [],
-            'aria-invalid': ariaInvalid,
-            onChange,
-            children,
-            ...otherProps
-        } = props
+export const InsightQueryInput: React.FunctionComponent<InsightQueryInputProps> = ({
+    value,
+    patternType,
+    repoQuery,
+    repositories = [],
+    'aria-invalid': ariaInvalid,
+    onChange,
+    children,
+    className,
+    ...otherProps
+}) => {
+    const repoQueryPreview =
+        repoQuery !== null ? getRepoQueryPreview(repoQuery) : generateRepoFiltersQuery(repositories)
+    const previewQuery = `${repoQueryPreview} ${value}`.trim()
 
-        const repoQueryPreview =
-            repoQuery !== null ? getRepoQueryPreview(repoQuery) : generateRepoFiltersQuery(repositories)
-        const previewQuery = `${repoQueryPreview} ${props.value}`.trim()
-
-        const handleOnChange = (queryState: QueryState): void => {
-            if (queryState.query !== value) {
-                onChange(queryState.query)
-            }
+    const handleOnChange = (queryState: QueryState): void => {
+        if (queryState.query !== value) {
+            onChange(queryState.query)
         }
+    }
 
-        return (
-            <div className={styles.root}>
-                {children ? (
-                    <Monaco.Root className={classNames(props.className, styles.inputWrapper)}>
-                        <Monaco.Field
-                            {...otherProps}
-                            ref={reference}
-                            patternType={patternType}
-                            queryState={{ query: value, changeSource: QueryChangeSource.userInput }}
-                            className={props.className}
-                            onChange={handleOnChange}
-                        />
-
-                        {children}
-                    </Monaco.Root>
-                ) : (
-                    <Monaco.Field
+    return (
+        <div className={styles.root}>
+            {children ? (
+                <FocusContainer className={classNames(className, styles.inputWrapper)}>
+                    <Field
                         {...otherProps}
-                        ref={reference}
                         patternType={patternType}
                         queryState={{ query: value, changeSource: QueryChangeSource.userInput }}
-                        className={classNames(styles.inputWrapper, props.className)}
+                        className={className}
                         onChange={handleOnChange}
                     />
-                )}
 
-                <Monaco.PreviewLink query={previewQuery} patternType={patternType} className={styles.previewButton}>
-                    Preview results <LinkExternalIcon size={18} />
-                </Monaco.PreviewLink>
-            </div>
-        )
-    }
-)
+                    {children}
+                </FocusContainer>
+            ) : (
+                <Field
+                    {...otherProps}
+                    patternType={patternType}
+                    queryState={{ query: value, changeSource: QueryChangeSource.userInput }}
+                    className={classNames(styles.inputWrapper, className)}
+                    onChange={handleOnChange}
+                />
+            )}
+
+            <PreviewLink query={previewQuery} patternType={patternType} className={styles.previewButton}>
+                Preview results <LinkExternalIcon size={18} />
+            </PreviewLink>
+        </div>
+    )
+}
